\name{antTraits.RData}
\alias{antTraits.RData}
\docType{data}
\title{
Ant traits data (this file needs updating)
}
\description{
Ant abundances for 60 species across 20 sites in eastern Australia, with species trait and environmental data 
}
\usage{data("antTraits.RData")}
\format{
  The format is:
 chr "antTraits.RData"
}
\details{
Details coming soon...
}
\source{
Coming soon...
}
\examples{
data(antTraits.RData)
## insert something here
}
\keyword{datasets}
