% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.manyglm.R
\name{coefplot.manyglm}
\alias{coefplot.manyglm}
\title{Plots the coefficients of the covariates of a manyglm object with confidence intervals.}
\usage{
coefplot.manyglm(object, y.label = TRUE, which.Xcoef = NULL,
  which.Ys = NULL, incl.intercept = FALSE, cex.ylab = 0.5, mfrow = NULL,
  mar = c(4, 8, 2, 1), ...)
}
\arguments{
\item{object}{A manyglm object}

\item{y.label}{Whether all the Y variables should be labelled}

\item{which.Xcoef}{Which X covariates should be included in the plot.}

\item{which.Ys}{Which Y variables should be included in the plot.}

\item{incl.intercept}{Whether the intercept coefficient should be included.}

\item{cex.ylab}{A plotting parameter. The default is 0.5.}

\item{mfrow}{Plotting parameter}

\item{mar}{Plotting parameter}

\item{...}{Other plotting parameters}
}
\value{
none
}
\description{
A way to plot the coefficients of the covariates of a manyglm object. Modifies code from Niku, Hui and Taskinen's coefplot.gllvm
}
