\name{mvabund-internal}
\title{Internal mvabund Objects}
\alias{LLCalc}
\alias{LLSolve}
\alias{[.mvabund}
\alias{absspp}
%\alias{anova.manyglmlist}
%\alias{anova.manylmlist}
\alias{as.data.frame.mvabund}
\alias{as.matrix.mvabund}
\alias{betaest}
%\alias{betaphiest}
%\alias{case.names.manyglm}
\alias{case.names.manylm}
\alias{cbind.mvabund}
% covratio: this is for the original function that is made generic to avoid warning while making the package
\alias{covratio} 
% dffits: this is for the original function that is made generic to avoid warning while making the package
\alias{dffits}
\alias{effic.make.link}
%\alias{formulaMva}
%\alias{geeSummaryTest}
%\alias{geeTest}
%\alias{geecalc}
%\alias{geecalcI}
%\alias{geecalcLR}
%\alias{geecalcsumm}
%\alias{geecalcsummI}
%\alias{geecalcsummLR}
%\alias{get.colors}
\alias{influence.measures}
\alias{is.formulaUnimva}
%\alias{makePlot}
\alias{manylm.multiwfit}
\alias{nbinall}
\alias{nbinall2}
\alias{nbinfit}
% plot: this is for the original function that is made generic to avoid warning while making the package
\alias{plot}
\alias{plotFormulafeature}
%\alias{print.anova.manyglm}
\alias{print.mvformula}
\alias{print.manyglm}
\alias{print.manylm}
\alias{print.mvabund}
\alias{rbind.mvabund}
\alias{rcalc}
% \alias{shiftpoints}
\alias{transAxis}
%\alias{variable.names.manyglm}
% weighted.residuals: this is for the original function that is made generic to avoid warning while making the package
%\alias{weighted.residuals}
%\alias{weights.manyglm}
\alias{RcppVersion}
\description{Internal mvabund functions.}
\details{These are not to be called by the user.}
\keyword{internal}

