\name{antTraits}
\docType{data}
\alias{antTraits}
% \encoding{latin1}
\title{Ant data, with species traits}
\description{
Ant abundances for 60 species across 20 sites in eastern Australia, with species trait and environmental data 
}
\usage{
data(antTraits)
}
\format{
A list containing three elements:
\describe{
\item{abund}{A data frame with 20 observations at different locations
 of abundances on 60 species.
}
\item{env}{A data frame containing 7 environmental variables measured at each of the 20 sites.}
\item{traits}{A data frame containing 6 species traits measured for each of the 60 species.}
}
}
\details{
Details coming soon...
}
\source{
Coming soon...
}
\examples{
data(antTraits)
## insert something here
}
\keyword{datasets}
